// Facetech.eMotoring/Presentation/Applications/Facetech.UI.Ngx.Nx.eMotoring/libs/common-ui/src/lib/person/models/interfaces/person.interface.ts
export interface IPersonNext {
  firstname: string;
  lastname: string;
  birthDate: string;
  baseActionValue: string;
  baseFunctionValue: string;
  baseSecuritySessionId: string;
  baseSessionValue: string;
  baseRequiredDatabase: string[];
  photos: string[];
  fingerprints?: string[];
  signatures?: string[];
  documents?: string[];
}